requirements.txt
#required libraries
aiohttp==3.6.2
asyncio==3.4.3
Pillow==7.1.1
python-multipart==0.0.5
starlette==0.13.3
uvicorn==0.11.3
scikit-image==0.16.2
#keras==2.3.1
#keras-preprocessing==1.1.0
numpy==1.18.0
tensorflow==2.3.1

